fix: implement global species catalog context to prevent multiple instances
